Inserito file di utility lib con tutte le funzioni per encdenc
